requirements.txt
fastapi==0.112.1
python-multipart==0.0.5
mysql-connector-python==8.4.0
passlib==1.7.4
pydantic==1.10.7
PyJWT==2.9.0
python-dotenv==1.0.1
cryptography>=3.2
bcrypt>=3.2.0
pytz==2023.3.post1
uvicorn==0.30.6
requests==2.31.0